+++ doc/changelog.md
@@ -4,6 +4,7 @@
 
 Changes:
 
+- ([#38](../../../issues/38)) Allow definition of restart policy on container start
 - ([#36](../../../issues/36)) Change docker client dependency to non-shaded
 - ([#34](../../../issues/34)) Expose underlying Docker API ContainerInfo to rule clients
 - ([#35](../../../issues/35)) Allow defining custom container startup conditions.